Renamed module 'bl' to 'utils'
