+++ src/main/java/lt/vu/mif/ui/helpers/implementations/ProductHelper.java
@@ -4,8 +4,6 @@ import java.io.InputStream;
 import java.util.List;
 import java.util.Objects;
 import javax.transaction.Transactional;
-import lt.vu.mif.bl.interfaces.IProductParser;
-import lt.vu.mif.bl.search.ProductSearch;
 import lt.vu.mif.excel.ExcelProduct;
 import lt.vu.mif.excel.ProductExcelReader;
 import lt.vu.mif.model.product.Product;
@@ -17,6 +15,8 @@ import lt.vu.mif.ui.mappers.implementations.ProductMapper;
 import lt.vu.mif.ui.view.BoughtProductView;
 import lt.vu.mif.ui.view.CartProductView;
 import lt.vu.mif.ui.view.ProductView;
+import lt.vu.mif.utils.interfaces.IProductParser;
+import lt.vu.mif.utils.search.ProductSearch;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Component;
