+++ src/main/java/lt/vu/mif/ui/helpers/implementations/UserHelper.java
@@ -2,22 +2,20 @@ package lt.vu.mif.ui.helpers.implementations;
 
 import java.time.LocalDateTime;
 import java.util.List;
-
 import lt.vu.mif.authentication.UserService;
-import lt.vu.mif.bl.interfaces.ITokenGenerator;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
-
+import lt.vu.mif.email.interfaces.IEmailContentGenerator;
 import lt.vu.mif.email.interfaces.IEmailProvider;
 import lt.vu.mif.model.user.Role;
 import lt.vu.mif.model.user.User;
 import lt.vu.mif.model.user.UserTokenTuple;
 import lt.vu.mif.repository.repository.interfaces.IUserRepository;
-import lt.vu.mif.email.interfaces.IEmailContentGenerator;
 import lt.vu.mif.ui.helpers.interfaces.IUserHelper;
 import lt.vu.mif.ui.mappers.implementations.UserMapper;
 import lt.vu.mif.ui.view.UserView;
+import lt.vu.mif.utils.interfaces.ITokenGenerator;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
 
 @Component
 public class UserHelper implements IUserHelper {
