+++ src/main/java/lt/vu/mif/ui/validation/PaymentCardValidation.java
@@ -7,9 +7,8 @@ import javax.faces.context.FacesContext;
 import javax.faces.validator.Validator;
 import javax.faces.validator.ValidatorException;
 import javax.inject.Named;
+import lt.vu.mif.utils.validation.ValidationUtils;
 import org.springframework.stereotype.Component;
-
-import lt.vu.mif.bl.validation.ValidationUtils;
 
 @Named
 @Component
