+++ src/main/java/lt/vu/mif/utils/constants/Constants.java
@@ -1,4 +1,4 @@
-package lt.vu.mif.bl.constants;
+package lt.vu.mif.utils.constants;
 
 public class Constants {
     public static final String PASSWORD_REQUEST_PARAMETER = "token";
