+++ src/main/java/lt/vu/mif/utils/interfaces/IImageDownloader.java
@@ -1,4 +1,4 @@
-package lt.vu.mif.bl.interfaces;
+package lt.vu.mif.utils.interfaces;
 
 public interface IImageDownloader {
     byte[] downloadImage(String imageLink);
