+++ src/main/java/lt/vu/mif/utils/interfaces/ITokenGenerator.java
@@ -1,4 +1,4 @@
-package lt.vu.mif.bl.interfaces;
+package lt.vu.mif.utils.interfaces;
 
 public interface ITokenGenerator {
     String generateToken();
